refactor(wcag-techniques): `WCAGTechniques` is now `WCAGTechniquesRunner`, `WCAGTechniquesModule` is now `WCAGTechniques`
This should better align meaning with readability of external API.
